chore: colocate endpoint-related types
