+++ src/services/transfer/types.ts
@@ -1,6 +1,4 @@
 import { AuthorizationRequirementsError } from '../../core/errors.js';
-import { ExtractKeys, Segment } from '../types.js';
-import { EndpointDocument } from './service/endpoint.js';
 
 /**
  * @see https://docs.globus.org/api/transfer/overview/#errors
@@ -24,116 +22,6 @@ export type CommonQueryParameters = {
   fields?: string;
   filter?: string;
 };
-
-export type EntityType =
-  | 'GCSv5_endpoint'
-  | 'GCSv5_mapped_collection'
-  | 'GCSv5_guest_collection'
-  | 'GCP_mapped_collection'
-  | 'GCP_guest_collection';
-
-/**
- * @see https://docs.globus.org/api/transfer/gcp_management/#update_collection_by_id
- */
-type GuestCollectionUpdatableField = ExtractKeys<
-  EndpointDocument,
-  | 'acl_max_expiration_period_mins'
-  | 'display_name'
-  | 'organization'
-  | 'department'
-  | 'keywords'
-  | 'description'
-  | 'contact_email'
-  | 'contact_info'
-  | 'info_link'
-  | 'owner_string'
-  | 'default_directory'
-  | 'force_encryption'
-  | 'disable_verify'
->;
-
-/**
- * @see https://docs.globus.org/api/transfer/gcp_management/#update_collection_by_id
- */
-type MappedCollectionUpdatableField = ExtractKeys<
-  EndpointDocument,
-  | GuestCollectionUpdatableField
-  | 'authentication_timeout_mins'
-  | 'subscription_id'
-  | 'public'
-  | 'location'
-  | 'network_use'
-  | 'max_concurrency'
-  | 'preferred_concurrency'
-  | 'max_parallelism'
-  | 'preferred_parallelism'
-  | 'user_message'
-  | 'user_message_link'
->;
-
-/**
- * @see https://docs.globus.org/api/transfer/endpoint_roles/
- */
-export type EndpointRole =
-  | 'administrator'
-  | 'access_manager'
-  | 'activity_manager'
-  | 'activity_monitor';
-
-/**
- * Mimics the OpenAPI generated types for Transfer (does not provide OpenAPI spec).
- *
- * The usage (and expansion) of this interface is deprecated. Types should be defined, and exported, in the
- * service methods that use them, and should not be defined here.
- *
- * @deprecated Use service method co-located types instead.
- * @private
- */
-export interface Operations
-  extends Record<
-    string,
-    {
-      parameters?: Segment;
-      response?: any;
-      payload?: any;
-    }
-  > {
-  CreateEndpoint: {
-    payload: Partial<Pick<EndpointDocument, GuestCollectionUpdatableField>> & {
-      DATA_TYPE?: 'shared_endpoint';
-      host_endpoint_id: string;
-      host_path: string;
-    };
-    response: {
-      DATA_TYPE: 'endpoint_create_result';
-      code: 'Created';
-      globus_connect_setup_key: string | null;
-      id: string;
-      message: string;
-      request_id: string;
-      resource: '/shared_endpoint';
-    };
-  };
-  UpdateEndpoint: {
-    payload: Partial<Pick<EndpointDocument, MappedCollectionUpdatableField | 'DATA_TYPE'>>;
-    response: {
-      DATA_TYPE: 'result';
-      code: 'Updated';
-      message: string;
-      request_id: string;
-      resource: `/endpoint/${string}`;
-    };
-  };
-  RemoveEndpoint: {
-    response: {
-      DATA_TYPE: 'result';
-      code: 'Deleted';
-      message: string;
-      request_id: string;
-      resource: `/endpoint/${string}`;
-    };
-  };
-}
 
 /**
  * Used to define query parameter types for the Transfer service.
